extract CodeBlockPart
